Update: combine resolve and inflate streams
