+++ package.json
@@ -12,7 +12,8 @@
     "posttest": "npm run lint && npm run check-commit",
     "coverage": "istanbul report --reporter=html && open coverage/lcov-report/index.html",
     "codecov": "cat coverage/coverage-final.json | codecov",
-    "compile": "babel src --out-dir lib --source-maps && pegjs src/transclude.pegjs lib/transclude-parser.js && pegjs src/inflater/link.pegjs lib/inflater/link.js",
+    "compile": "babel src --out-dir lib --source-maps && npm run compile-grammars",
+    "compile-grammars": "mkdir -p ./lib/grammar && pegjs src/grammar/transclusion-link.pegjs lib/grammar/transclusion-link.js && pegjs src/grammar/inflate-link.pegjs lib/grammar/inflate-link.js",
     "lint": "eslint ./src ./test",
     "install-bats": "./scripts/install-bats",
     "check-commit": "./scripts/commit-msg",
